Upgraded js-xlsx, plotly and js-beautify (bower.json) #993

+++ src/main/webapp/bower_components/js-xlsx/xlsxworker1.js
@@ -9,8 +9,8 @@ postMessage({t:"ready"});
 
 function ab2str(data) {
 	var o = "", l = 0, w = 10240;
-	for(; l<data.byteLength/w; ++l) o+=String.fromCharCode.apply(null,new Uint16Array(data.slice(l*w,l*w+w)));
-	o+=String.fromCharCode.apply(null, new Uint16Array(data.slice(l*w)));
+	for(; l<data.byteLength/w; ++l) o+=String.fromCharCode.apply(null,new Uint8Array(data.slice(l*w,l*w+w)));
+	o+=String.fromCharCode.apply(null, new Uint8Array(data.slice(l*w)));
 	return o;
 }
